requirements.txt
speedtest-cli==2.1.3
requests==2.27.1
python-telegram-bot==13.7